Move methods up to VegaFusionRuntimeTrait
