+++ vegafusion-core/src/runtime/mod.rs
@@ -1,5 +1,5 @@
 mod runtime;
-pub use runtime::{PreTransformExtractTable, VegaFusionRuntimeTrait};
+pub use runtime::{PreTransformExtractTable, VegaFusionRuntimeTrait, decode_inline_datasets};
 
 #[cfg(feature = "tonic_support")]
 mod grpc_runtime;
