+++ vegafusion-server/src/main.rs
@@ -17,7 +17,7 @@ use vegafusion_core::proto::gen::tasks::{
     ResponseTaskValue,
     TaskValue as ProtoTaskValue,
 };
-use vegafusion_core::runtime::VegaFusionRuntimeTrait;
+use vegafusion_core::runtime::{VegaFusionRuntimeTrait, decode_inline_datasets};
 use vegafusion_core::spec::chart::ChartSpec;
 use vegafusion_core::task_graph::graph::ScopedVariable;
 use vegafusion_runtime::task_graph::runtime::VegaFusionRuntime;
@@ -50,7 +50,7 @@ impl VegaFusionRuntimeGrpc {
                 let task_graph = Arc::new(task_graph_values.task_graph.unwrap());
                 let indices = &task_graph_values.indices;
                 let inline_datasets =
-                    VegaFusionRuntime::decode_inline_datasets(task_graph_values.inline_datasets)?;
+                    decode_inline_datasets(task_graph_values.inline_datasets)?;
 
                 match self
                     .runtime
@@ -60,7 +60,9 @@ impl VegaFusionRuntimeGrpc {
                     Ok(response_values) => {
                         let response_msg = QueryResult {
                             response: Some(query_result::Response::TaskGraphValues(
-                                TaskGraphValueResponse { response_values },
+                                TaskGraphValueResponse { 
+                                    response_values: response_values.into_iter().map(|v| v.into()).collect::<Vec<_>>() 
+                                },
                             )),
                         };
                         Ok(response_msg)
@@ -97,7 +99,7 @@ impl VegaFusionRuntimeGrpc {
         });
 
         // Decode inline datasets to VegaFusionDatasets
-        let inline_datasets = VegaFusionRuntime::decode_inline_datasets(request.inline_datasets)?;
+        let inline_datasets = decode_inline_datasets(request.inline_datasets)?;
 
         // Parse spec
         let spec: ChartSpec = serde_json::from_str(&request.spec)?;
@@ -129,7 +131,7 @@ impl VegaFusionRuntimeGrpc {
         // Extract and deserialize inline datasets
         let inline_pretransform_datasets = request.inline_datasets;
         let inline_datasets =
-            VegaFusionRuntime::decode_inline_datasets(inline_pretransform_datasets)?;
+            decode_inline_datasets(inline_pretransform_datasets)?;
         let opts = request.opts.unwrap();
 
         // Parse spec
@@ -182,7 +184,7 @@ impl VegaFusionRuntimeGrpc {
             });
 
         // Extract and deserialize inline datasets
-        let inline_datasets = VegaFusionRuntime::decode_inline_datasets(request.inline_datasets)?;
+        let inline_datasets = decode_inline_datasets(request.inline_datasets)?;
 
         // Extract requested variables
         let variables: Vec<ScopedVariable> = request
